Add run config back to helper output
Using the helper within the asset pipeline yields different checksums for all the files
vs outputting the config on the page within the helper. This must be fixed in order to get the
require.js config into an external JS file so we are reverting to outputting the config
inline on the page at the expense of caching.